refactor: 도메인 생애주기별로 패키지 분리
