+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
 import kitchenpos.application.mapper.TableGroupMapper;
-import kitchenpos.domain.order.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.dto.request.TableGroupRequest;
 import kitchenpos.dto.response.TableGroupResponse;
 import kitchenpos.repository.TableGroupRepository;
