+++ src/main/java/kitchenpos/application/mapper/MenuMapper.java
@@ -2,11 +2,11 @@ package kitchenpos.application.mapper;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.domain.Price;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.menu.Price;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 import kitchenpos.dto.request.MenuProductRequest;
 import kitchenpos.dto.request.MenuRequest;
 import kitchenpos.repository.MenuGroupRepository;
