+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.menu;
+package kitchenpos.domain.product;
 
 import java.math.BigDecimal;
 import javax.persistence.Embedded;
@@ -7,6 +7,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import kitchenpos.domain.Price;
 
 @Table(name = "product")
 @Entity
