+++ src/main/java/kitchenpos/dto/request/OrderTableRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.request;
 
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 
 public class OrderTableRequest {
