+++ src/main/java/kitchenpos/dto/request/ProductRequest.java
@@ -1,7 +1,7 @@
 package kitchenpos.dto.request;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 
 public class ProductRequest {
