+++ src/main/java/kitchenpos/dto/response/ProductResponse.java
@@ -1,7 +1,7 @@
 package kitchenpos.dto.response;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 
 public class ProductResponse {
