+++ src/main/java/kitchenpos/dto/response/TableGroupResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.order.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 public class TableGroupResponse {
