+++ src/main/java/kitchenpos/repository/JpaOrderTableRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.repository.Repository;
 
 public interface JpaOrderTableRepository extends Repository<OrderTable, Long>, OrderTableRepository {
