+++ src/main/java/kitchenpos/repository/JpaProductRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 import org.springframework.data.repository.Repository;
 
 public interface JpaProductRepository extends Repository<Product, Long>, ProductRepository {
