+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 import kitchenpos.dto.request.MenuProductRequest;
 import kitchenpos.dto.request.MenuRequest;
 import kitchenpos.dto.response.MenuResponse;
