+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -5,12 +5,12 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.DatabaseCleaner;
+import kitchenpos.domain.Price;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.menu.Price;
-import kitchenpos.domain.menu.Product;
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.domain.product.Product;
 import kitchenpos.dto.request.MenuGroupRequest;
 import kitchenpos.dto.request.MenuProductRequest;
 import kitchenpos.dto.request.MenuRequest;
