+++ src/test/java/kitchenpos/domain/menu/ProductTest.java
@@ -3,6 +3,7 @@ package kitchenpos.domain.menu;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
+import kitchenpos.domain.product.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
