+++ src/test/java/kitchenpos/domain/order/OrderTableTest.java
@@ -2,6 +2,7 @@ package kitchenpos.domain.order;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import kitchenpos.domain.ordertable.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
